apps/api/pyproject.toml
[project]
name = "rootspread-api"
version = "0.1.0"
description = "FastAPI backend for RootSpread"
readme = "README.md"
authors = [
    { name = "Mark", email = "FB208@outlook.com" }
]
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "email-validator>=2.3.0",
    "fastapi>=0.135.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.13.1",
    "pymysql>=1.1.2",
    "python-jose[cryptography]>=3.5.0",
    "redis>=7.3.0",
    "resend>=2.23.0",
    "sqlalchemy>=2.0.48",
    "uvicorn[standard]>=0.41.0",
]

[project.scripts]
rootspread-api = "rootspread_api.main:run"

[build-system]
requires = ["uv_build>=0.9.13,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.5",
]

[tool.ruff]
line-length = 100

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
